+++ src/sidebar/theme.js
@@ -4,21 +4,19 @@ function getThemeFromStorage() {
   const getting = browser.storage.local.get(['theme']);
   getting.then(function applyTheme(data) {
     if (data.theme === 'dark') {
-      if (! document.getElementById('dark-styles')) {
+      if (!document.getElementById('dark-styles')) {
         const darkSS = document.createElement('link');
         darkSS.id = 'dark-styles';
         darkSS.type = 'text/css';
         darkSS.rel = 'stylesheet';
         darkSS.href = 'styles-dark.css';
         document.getElementsByTagName('head')[0].appendChild(darkSS);
-      } else
-        return;
+      }
     } else if (data.theme === 'default' || data.theme === undefined) {
       if (document.getElementById('dark-styles')) {
         const darkSS = document.getElementById('dark-styles');
         darkSS.parentElement.removeChild(darkSS);
-      } else
-        return;
+      }
     }
   });
 }